fixed positive label
The positive label is now calculated every time the target is changed and is binary.
